cleanup: splitted file to sliderWithTarget fixed slider refresh when changing slider step
