+++ src/yarpmotorgui-qt/jointitem.cpp
@@ -18,117 +18,6 @@
 #include <yarp/os/Log.h>
 #include <QDebug>
 #include <QKeyEvent>
-#include <QPainter>
-
-SliderWithTarget::SliderWithTarget(QWidget * parent, bool _hasTargetOption) : QSlider(Qt::Horizontal, 0)
-{
-    sliderLabel = new QLabel(this);
-    sliderLabel->setObjectName(QStringLiteral("sliderLabel"));
-    sliderLabel->setGeometry(QRect(0, 0, 41, 20));
-    sliderLabel->setMinimumSize(QSize(40, 20));
-    sliderLabel->setMaximumSize(QSize(16777215, 20));
-    sliderLabel->setAlignment(Qt::AlignCenter);
-    sliderLabel->setStyleSheet("background-color: rgba(0,0,0,0%)");
-
-    hasTargetOption = _hasTargetOption;
-    enableViewTarget = true;
-    enableViewLabel = true;
-    disableClickOutOfHandle = true;
-    target=-1e50; //default value is intensionally set out of the slider
-    width_at_target = this->width() - 30;
-    sliderStep=1;
-    isDouble = false;
-}
-
-void SliderWithTarget::setSliderStep(double step)
-{
-    sliderStep = step;
-}
-
-double SliderWithTarget::getSliderStep()
-{
-    return sliderStep;
-}
-
-void SliderWithTarget::setIsDouble(bool b)
-{
-    isDouble = b;
-}
-
-bool SliderWithTarget::getIsDouble()
-{
-    return isDouble;
-}
-
-void SliderWithTarget::setValue(double val)
-{
-    if (isDouble)
-    {
-        QSlider::setValue(round(val*sliderStep));
-    }
-    else
-    {
-        QSlider::setValue(val);
-    }
-}
-
-void SliderWithTarget::updateSliderTarget(double t)
-{
-    target = t;
-    width_at_target= this->width()-30;
-}
-
-void SliderWithTarget::resetTarget()
-{
-    target = target = -1e50;
-    width_at_target = this->width() - 30;
-}
-
-void SliderWithTarget::paintEvent(QPaintEvent *e)
-{
-    QSlider::paintEvent(e);
-    if (hasTargetOption && enableViewTarget)
-    {
-        QPainter p(this);
-        double current_width = this->width()-30;
-        QRect r(target / width_at_target * current_width, 17, 30, 15);
-        p.fillRect(r, QBrush(QColor(128, 128, 255, 128)));
-        p.drawRect(r);
-    }
-    if (enableViewLabel)
-    {
-        double  value = this->value();
-        int w = this->width() - 30;
-        int totValues = this->maximum() + abs(this->minimum());
-        double newX = ((double)w / (double)totValues) * ((double)value + abs(this->minimum()));
-        sliderLabel->setGeometry(newX, -10, 40, 20);
-        sliderLabel->setText(QString("%L1").arg((double)value / sliderStep, 0, 'f', 3));
-    }
-}
-
-void SliderWithTarget::mousePressEvent(QMouseEvent * event)
-{
-    if (event->button() == Qt::LeftButton)
-    {
-        double myx = event->x();
-        QStyleOptionSlider opt;
-        initStyleOption(&opt);
-        QRect sr = style()->subControlRect(QStyle::CC_Slider, &opt, QStyle::SC_SliderHandle, this);
-
-        //throw aways clicks that are not on the handle
-        if (disableClickOutOfHandle)
-        {
-            if (myx<sr.left()+3 || myx>sr.right()-3)
-            {
-                return;
-            }
-        }
-
-        //setValue(minimum() + ((maximum() - minimum()) * myx) / width());
-        event->accept();
-    }
-    QSlider::mousePressEvent(event);
-}
 
 void JointItem::resetTarget()
 {
@@ -727,6 +616,7 @@ void JointItem::enableVelocitySliderDoubleAuto()
     int v = ui->sliderVelocityVelocity->value();
     if (v > sliderMax) {}
     if (v < sliderMin) {}
+    setRefVelocitySpeed(ref_speed);
 }
 
 void JointItem::enableVelocitySliderDoubleValue(double value)
@@ -740,6 +630,7 @@ void JointItem::enableVelocitySliderDoubleValue(double value)
     int v = ui->sliderVelocityVelocity->value();
     if (v > sliderMax) {}
     if (v < sliderMin) {}
+    setRefVelocitySpeed(ref_speed);
 }
 
 void JointItem::disableVelocitySliderDouble()
@@ -758,6 +649,7 @@ void JointItem::disableVelocitySliderDouble()
     int v = ui->sliderVelocityVelocity->value();
     if (v > sliderMax) {}
     if (v < sliderMin) {}
+    setRefVelocitySpeed(ref_speed);
 }
 
 void JointItem::enableTorqueSliderDoubleAuto()
@@ -772,6 +664,7 @@ void JointItem::enableTorqueSliderDoubleAuto()
     int v = ui->sliderTorqueTorque->value();
     if (v > sliderMax) {}
     if (v < sliderMin) {}
+    setRefTorque(ref_torque);
 }
 
 void JointItem::enableTorqueSliderDoubleValue(double value)
@@ -785,6 +678,7 @@ void JointItem::enableTorqueSliderDoubleValue(double value)
     int v = ui->sliderTorqueTorque->value();
     if (v > sliderMax) {}
     if (v < sliderMin) {}
+    setRefTorque(ref_torque);
 }
 
 void JointItem::disableTorqueSliderDouble()
@@ -803,6 +697,7 @@ void JointItem::disableTorqueSliderDouble()
     int v = ui->sliderTorqueTorque->value();
     if (v > sliderMax) {}
     if (v < sliderMin) {}
+    setRefTorque(ref_torque);
 }
 
 void JointItem::enableTrajectoryVelocitySliderDoubleAuto()
@@ -817,6 +712,7 @@ void JointItem::enableTrajectoryVelocitySliderDoubleAuto()
     int v = ui->sliderTrajectoryVelocity->value();
     if (v > sliderMax) {}
     if (v < sliderMin) {}
+    setRefTrajectorySpeed(ref_trajectory_velocity);
 }
 
 void JointItem::enableTrajectoryVelocitySliderDoubleValue(double value)
@@ -830,6 +726,7 @@ void JointItem::enableTrajectoryVelocitySliderDoubleValue(double value)
     int v = ui->sliderTrajectoryVelocity->value();
     if (v > sliderMax) {}
     if (v < sliderMin) {}
+    setRefTrajectorySpeed(ref_trajectory_velocity);
 }
 
 void JointItem::disableTrajectoryVelocitySliderDouble()
@@ -843,6 +740,7 @@ void JointItem::disableTrajectoryVelocitySliderDouble()
     int v = ui->sliderTrajectoryVelocity->value();
     if (v > sliderMax) {}
     if (v < sliderMin) {}
+    setRefTrajectorySpeed(ref_trajectory_velocity);
 }
 
 void JointItem::setEnabledOptions(bool debug_param_enabled,
@@ -1260,6 +1158,7 @@ void JointItem::setRefTorque(double val)
 {
     if(ui->stackedWidget->currentIndex() == TORQUE){
         updateSliderTorque(val);
+        ref_torque = val;
     }
 }
 
@@ -1267,6 +1166,7 @@ void JointItem::setRefVelocitySpeed(double val)
 {
     if (ui->stackedWidget->currentIndex() == VELOCITY){
         updateSliderVelocity(val);
+        ref_speed = val;
     }
 }
 
@@ -1274,6 +1174,7 @@ void JointItem::setRefTrajectorySpeed(double val)
 {
     if (ui->stackedWidget->currentIndex() == POSITION){
         updateSliderTrajectoryVelocity(val);
+        ref_trajectory_velocity = val;
     }
 }
